Refactoring assoc_array logic
Using a base class to avoid cross-includes between the API and AST
directories.
