+++ lib/coek/coek/api/variable_array.cpp
@@ -1,9 +1,8 @@
-
 #include "coek/ast/value_terms.hpp"
+#include "coek/api/variable_array.hpp"
 #include "coek/api/variable_assoc_array_repn.hpp"
 #include "coek/model/model.hpp"
 #include "coek/model/model_repn.hpp"
-#include "coek/api/variable_array.hpp"
 
 namespace coek {
 
@@ -40,11 +39,11 @@ class VariableArrayRepn : public VariableAssocArrayRepn {
 
     virtual ~VariableArrayRepn() {}
 
-    Variable index(const IndexVector& args);
+    std::shared_ptr<VariableTerm> index(const IndexVector& args);
 
-    size_t dim() const { return shape.size(); }
+    size_t dim() { return shape.size(); }
 
-    size_t size() const { return _size; }
+    size_t size() { return _size; }
 
     std::string get_name(std::string name, size_t index);
 
@@ -80,7 +79,7 @@ void VariableArrayRepn::generate_names()
     // If no name has been provided to this array object,
     // then we do not try to generate names.  The default/simple
     // variable names will be used.
-    std::string name = variable_template.name();
+    std::string name = value_template.name();
     if (name == "")
         return;
 
@@ -127,7 +126,7 @@ const std::shared_ptr<VariableAssocArrayRepn> VariableArray::get_repn() const { 
 Variable VariableArray::index(const IndexVector& args)
 { return repn->index(args); }
 
-Variable VariableArrayRepn::index(const IndexVector& args)
+std::shared_ptr<VariableTerm> VariableArrayRepn::index(const IndexVector& args)
 {
     //auto _repn = repn.get();
     //auto& shape = _repn->shape;
@@ -142,7 +141,7 @@ Variable VariableArrayRepn::index(const IndexVector& args)
         ndx = ndx * shape[i] + static_cast<size_t>(args[i]);
 
     if (ndx > size()) {
-        std::string err = "Unknown index value: " + variable_template.name() + "[";
+        std::string err = "Unknown index value: " + value_template.name() + "[";
         for (size_t i = 0; i < args.size(); i++) {
             if (i > 0)
                 err += ",";
@@ -152,13 +151,13 @@ Variable VariableArrayRepn::index(const IndexVector& args)
         throw std::runtime_error(err);
     }
 
-    return values[ndx];
+    return values[ndx].repn;
 }
 
 void VariableArray::index_error(size_t i)
 {
     auto _repn = repn.get();
-    std::string err = "Unexpected index value: " + _repn->variable_template.name() + " is an "
+    std::string err = "Unexpected index value: " + _repn->value_template.name() + " is an "
                       + std::to_string(tmp.size()) + "-D variable array but is being indexed with "
                       + std::to_string(i) + " indices.";
     throw std::runtime_error(err);
@@ -254,12 +253,12 @@ VariableArray& VariableArray::fixed(bool value)
 
 VariableArray& VariableArray::name(const std::string& name)
 {
-    repn->variable_template.name(name);
+    repn->value_template.name(name);
     repn->name(name);
     return *this;
 }
 
-std::string VariableArray::name() const { return repn->variable_template.repn->name; }
+std::string VariableArray::name() const { return repn->value_template.repn->name; }
 
 VariableArray& VariableArray::within(VariableTypes vtype)
 {
